Bump dependencies and remove unused ones
Manually checked that each bumped dependency doesn't break anything we depend on.

This means bumping the minimum version of Node.js to 0.12 as gulp-autoprefixer and gulp-size requires Node.js 0.12. Doesn't matter though as this is used in a dev setting, so newer Node.js should be expected. This will just get worse, so no point in delaying it.