fix(list): ensure sass mixin works if leading is provided in px
This switches the subheader margin calculation from static to dynamic so
that it will work if the $mdc-list-subheader-leading is provided in
px instead of rem. Angular Material's current typography settings
specify it in px. We do plan to move away from that and use rem like
MDC, but this fix will make it work in the mean time.

PiperOrigin-RevId: 297181874