fix(mdc-slider): avoid server side rendering error
Recently with 3802a49b71045a189ec46197c54079ba86c34419, support for
pointer events has been added to the `MdcSlider`. The feature detection
for pointer events currently fails on the server as it accesses `window`
directly. We should not error if we are on the server as these events
are not having any effect anyway.