chore: Release
This PR was opened by [release][1] GitHub action.
When you're ready to do a release, you can merge this and release
action will automatically publishes to npm registry.
If you're not ready to do a release yet, that's fine, whenever you
add more commits to master, this PR will be updated.

[1]: http://github.com/material-components/material-components-web/tree/master/.github/workflows/release-pull-request.yml

PiperOrigin-RevId: 317874587