fix(mdc-list): invalid syntax in generated .d.ts bundle
An earlier commit added a doc string with an `@return` in it (43e8390de14e6f92e02a9a6a4b1e49f80dee959b). By itself this isn't wrong, but it seems to hit a buggy code path in `dts-bundle` which causes it not to close a comment and to produce an invalid .d.ts file.

These changes work around the issue by using `@returns` instead. A longer term solution would be to move away from `dts-bundle` since it doesn't appear to be maintained.