feat(mdc-list): introduce selection change event
Introduces a new custom event called `MDCList:selectionChange` that
fires whenever the selection of the list has changed through user
interaction. This matches with the semantics of the native `change`
event of input controls.

This event is useful as not every type of list has an underlying
native control, like in the checkbox-type list, where the `change`
event from the control may bubble up. Even for those cases though
the native event is not guaranteed to fire in all cases. e.g. it
does not fire when a user leverages the `CTRL + A` keybinding we
support in the list foundation.

This commit adds a new type of event/notification that can be
leveraged in the foundation through an adapter method (needed by
Angular), or consumers could use it through `MDCList:selectionChange`.

Note that the custom action event being emitted here is not semantically
equivalent as this one fires regardless of whether the selection has actually changed
or not. This seems like another use-cases, as seen in the MDC menu.

**Note**: also fixes that the unit tests did not run properly as there was a `fit` leftover pushed to `master`. This commit fixes that and sets up a lint to prevent this in the future.

PiperOrigin-RevId: 415565874