fix(snackbar): address Trusted Types violation
Assigning a string directly to the innerHTML property is potentially security sensitive, and thus causes a Trusted Types violation. This blocks apps using mdc-snackbar from adopting Trusted Types.

Refactor the innerHTML assignment to the equivalent construction using native DOM APIs which is compatible with Trusted Types.

This was tested on JAWS 2022/Chrome 99, verifying that screen reading behaviour was equivalent before and after this change.

PiperOrigin-RevId: 435754345