style(tooltip): Fixing tooltip's line-clamp mixin.
When line-clamp is used on an element with bottom-padding, the cut-off line of text is still visible (due to this bottom padding). The fix is to apply the padding on the parent element. Due to how tooltip is structured (the tooltip surface element applies the background color), we actually need to create an additional element around the label itself in order to maintain the expected styling.

Also updates the documentation to reflect the new structure of the tooltip (note that even if this new structure is adopted, the tooltip will continue to function as expected. The only thing that won't work is the `line-clamp` mixin.

PiperOrigin-RevId: 462231748