fix(tooltip): Fix incorrect measurements of rich tooltip widths on subsequent renders.
When rich tooltips are positioned within their parent containers, the tooltip width might be shrunk if it collides with the edge of the parent container. The width of the tooltip is set as a style property in order to prevent this. However, when rendered with short contents (e.g. single words), the returned `computedStyleProperty('width')` is slightly smaller than the originally computed width (e.g. `76.953px` vs `76.6875px`) which results in the content being split into multiple lines.

To prevent this, we "reset" the tooltip so the width measurement is accurate (this includes clearing the previously calculated width and moving the tooltip off to the left so the parent container doesn't restrict the width).

PiperOrigin-RevId: 474140842