feat(linearprogress): Implement separate `track-height` and `active-indicator-height` tokens
This commit updates `track-height` to only apply to the track and implements `active-indicator-height` that only applies to the bar (aka active indicator). Historically, Linear Progress's `track-height` token applied to both the bar and track. A previous commit added `bar-and-track-height` that applies to both bar and track, keeping existing functionality intact.

PiperOrigin-RevId: 529770168