fix(formfield): Change cursor style on label to `inherit` rather than `auto` to adopt customization.
radio button labels are interactive elements that users can click to select an option. To enhance the user experience and maintain uniformity across the interface, developers may want to customize the cursor appearance when the mouse hovers over the label. By setting the cursor style to `inherit`, developers gain the flexibility to modify the cursor style as needed, ensuring a consistent and visually appealing user interface.

PiperOrigin-RevId: 630025696