Removing junk
Stripping out the release stuff as I intend to vastly decrease what is needed to build everything if possible and this is going to immediately be old/wrong.

Removed all the crap related to code community standards etc as the point here is for the community to drive this not to be locked in enteprise standard bs. If it's useful submit a PR. Don't be a jerk. The end.

Removing all the ridiculous proccess about all the stuff you need to do to contribute for similar reason. It's what led to the death of this project partly imo. Too costly for Google to maintain.